refactor: extract lattice constructor methods to Lattice
