+++ qiskit_nature/second_q/hamiltonians/lattice_model.py
@@ -13,7 +13,6 @@
 """The Lattice Model class."""
 
 import numpy as np
-from retworkx import PyGraph
 
 from .hamiltonian import Hamiltonian
 from .lattices import Lattice
@@ -41,42 +40,3 @@ class LatticeModel(Hamiltonian):
             The interaction matrix.
         """
         return self._lattice.to_adjacency_matrix(weighted=True)
-
-    @staticmethod
-    def _generate_lattice_from_uniform_parameters(
-        lattice: Lattice,
-        uniform_interaction: complex,
-        uniform_onsite_potential: complex,
-    ) -> Lattice:
-        graph = lattice.graph
-        for node_a, node_b, _ in graph.weighted_edge_list():
-            if node_a != node_b:
-                graph.update_edge(node_a, node_b, uniform_interaction)
-
-        for node_a in graph.node_indexes():
-            if graph.has_edge(node_a, node_a):
-                graph.update_edge(node_a, node_a, uniform_onsite_potential)
-            else:
-                graph.add_edge(node_a, node_a, uniform_onsite_potential)
-
-        return Lattice(graph)
-
-    @staticmethod
-    def _generate_lattice_from_parameters(interaction_matrix: np.ndarray):
-        # make a graph from the interaction matrix.
-        # This should be replaced by from_adjacency_matrix of retworkx.
-        shape = interaction_matrix.shape
-        if len(shape) != 2 or shape[0] != shape[1]:
-            raise ValueError(
-                f"Invalid shape of `interaction_matrix`, {shape},  is given."
-                "It must be a square matrix."
-            )
-
-        graph = PyGraph(multigraph=False)
-        graph.add_nodes_from(range(shape[0]))
-        for source_index in range(shape[0]):
-            for target_index in range(source_index, shape[0]):
-                weight = interaction_matrix[source_index, target_index]
-                if not weight == 0.0:
-                    graph.add_edge(source_index, target_index, weight)
-        return Lattice(graph)
